remove polarity when converting inftype into type
originally this was thought to maybe be useful to get
more general closure types, but it turns out that this
makes them too general (for example, the identity closure
  <x> x
would infer as any(int) when apllied with an integer even
when it could be int(int) which would be much more useful.
When applying the bounds the opposite way to make the
type more specific, one gets the unsound type int(any) for
the closure. We could always choose the upper bound and
would get any(any) in this instance, however this is not useful
and is an issue with InferField, which cannot be lowered to a
concrete type and only appears in upper bounds. We do
the intuitive thing and use the lower bounds, which results
in int(int) in this case.