change pointer offset for multilayouts
when a function reads a value from an argument, before this it needed to
know whether the value was a multi-layout or not, since the data is
after the optional vtable pointer.

if the value is written into a multilayout place, but it is statically
known that the value is a monolayout, by putting the vtable pointer at
a negative index in front of the data, it is possible to use the value
as if it was a monolayout. the data always starts at offset 0, and at
-word one can optionally access the vtable pointer if needed.