change name mangling scheme to include from layout
the slot is not inherent information but just a derived value
that made the symbol unique. instead, we now just use a hash of the
from arg layout to make it unique so that we do not have to carry the
slot number everywhere.

this also makes it so that functions don't have to have a slot, which
is only really required for dynamic dispatch