yabo: fix bug where `else` with two parsers failed
the reason for this is that the vtable for the place was not initialized
as the zst was just assumed to need no initialization. this was fixed
by creating an extra stack place which will be mono and therefore
needs no initialization. the copy to the original place will then
initialize the vtable.