fix len handling
previously, the length checking was done somewhat
badly: Dynamic values represented multiple possible values,
and a parser that was unsized was considered as Dynamic
because for different inputs it could have different lengths.

this commits fixes that, and also the dependency handling,
which was lowered incorrectly, for example when a function
didn't need a particular argument but there was the possibility
of backtracking within the expression in the argument.